statisfund_replica/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
backtrader==1.9.78.123
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.2
openai==1.86.0
python-multipart==0.0.6
python-dotenv==1.0.0
